Navbar now shared across all pages
The navbar is now shared across all of the pages without needing to copy and paste the raw navbar code to each pages. This approach is fantastic because any changes I make to the navbar will update to all pages. This was achieved through using the .get property of jQuery. This function is able to get the data from the navbar.html, which contains all of the navbar code, and replaces the element with the id nav-placeholder with this data. Very convenient!